requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cleanup==8.0.0
django-environ==0.10.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-split-settings==1.2.0
Pillow==10.0.0
psycopg2==2.9.7
six==1.16.0
sqlparse==0.4.4
transliterate==1.10.2
typing_extensions==4.7.1
tzdata==2023.3
